+++ src/main/java/com/getpcpanel/cpp/linux/LinuxProcessHelper.java
@@ -9,13 +9,22 @@ import org.apache.commons.lang3.math.NumberUtils;
 import org.springframework.stereotype.Service;
 
 import com.getpcpanel.spring.ConditionalOnLinux;
+import com.getpcpanel.util.ProcessHelper;
 
+import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 
 @Log4j2
 @Service
 @ConditionalOnLinux
-public class ProcessHelper {
+@RequiredArgsConstructor
+public class LinuxProcessHelper {
+    private final ProcessHelper processHelper;
+
+    public ProcessBuilder builder(String... command) {
+        return processHelper.builder(command);
+    }
+
     public int getActiveProcessPid() {
         try {
             var activeWindow = lineFrom("xdotool", "getactivewindow");
@@ -42,7 +51,7 @@ public class ProcessHelper {
     }
 
     private String lineFrom(String... cmd) throws IOException {
-        var lines = IOUtils.readLines(new ProcessBuilder(cmd).start().getInputStream(), Charset.defaultCharset());
+        var lines = IOUtils.readLines(processHelper.builder(cmd).start().getInputStream(), Charset.defaultCharset());
         if (lines.isEmpty()) {
             return null;
         }