+++ src/main/java/com/getpcpanel/util/ProcessHelper.java
@@ -1,0 +1,12 @@
+package com.getpcpanel.util;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class ProcessHelper {
+    public ProcessBuilder builder(String... command) {
+        var result = new ProcessBuilder(command);
+        result.environment().put("LC_ALL", "C");
+        return result;
+    }
+}